% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_aa_distribution.R
\name{find_aa_distribution}
\alias{find_aa_distribution}
\title{Find the distribution of amino acids in the polypeptide.}
\usage{
find_aa_distribution(polypep)
}
\arguments{
\item{polypep}{}
}
\value{
polypptide_plot
}
\description{
Create a histogram of the distribution
}
\examples{
find_aa_distribution("HEGGHNRSRSSRTTSLKEET")
}
\author{
Miss Oriade Latifah Simpson, \email{s172084@dtu.dk}
}
