% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{the_codon_table}
\alias{the_codon_table}
\title{The Standard Code}
\format{
An dataframe
}
\source{
\href{https://www.ncbi.nlm.nih.gov/Taxonomy/Utils/wprintgc.cgi?chapter=tgencodes#SG1}{The Genetic Codes}
}
\usage{
data(the_codon_table)
}
\description{
Translation Table 1 from Genetic Codes
}
\examples{
data(the_codon_table)
amino_acid_sq <- paste0(the_codon_table[codons], collapse = "")
}
\references{
(Takahashi et al. (2005). Genomics 2005 Mar;85(3):360-71.  doi: 10.1016/j.ygeno.2004.11.012.
(\href{https://pubmed.ncbi.nlm.nih.gov/15718103/}{PubMed})
}
\keyword{datasets}
